+++ test_cmd.py
@@ -251,7 +251,7 @@ class TestCmd(unittest.TestCase):
           '(?: '
             r'[<] (?P<angle_bracketed_uri> [^>]*? ) [>]'
               ' | '
-            r'(?P<bare_uri> [\S]+ )'
+            r'(?P<bare_uri> [\S]+? )'
           ' )'
         ' )?'
         fr'(?: [\s]* '
@@ -278,7 +278,7 @@ class TestCmd(unittest.TestCase):
           '(?: '
             r'[<] (?P<angle_bracketed_uri> [^>]*? ) [>]'
               ' | '
-            r'(?P<bare_uri> [\S]+ )'
+            r'(?P<bare_uri> [\S]+? )'
           ' )'
         ' )?'
         fr'(?: [\s]* '
@@ -1518,8 +1518,8 @@ span across blocks]<a href="file.svg" title="title" class="test">label</a>
 <a href="spacious/href" title="spacious title">Spacious text.</a>
 <a href="H" title="T">Link</a>
 <a href="href" title="title2"><img alt="alt" src="src" title="title"></a>
-In brackets href: (<a href="href)">text</a>
-In brackets title: (<a href="&quot;title&quot;)">text</a>
+In brackets href: (<a href="href">text</a>)
+In brackets title: (<a href="&quot;title&quot;">text</a>)
 In brackets both: (<a href="href" title="title">text</a>)
 <h2><code>#referenced-links</code></h2>
 <a href="/file" title="title" class="test">Content.</a>